backend/requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# LangChain (text splitting, RAG chain helpers)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

# Vector database
chromadb>=0.5.0

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
